// src/integrations/fetch/scraper-type.ts
export enum ScraperType {
  JINA_READER = "JINA_READER",
  JINA_SEARCH = "JINA_SEARCH",
  JINA_DEEPSEARCH = "JINA_DEEPSEARCH",
  BRAVE_SEARCH = "BRAVE_SEARCH",
  TAVILY_SEARCH = "TAVILY_SEARCH",
  EXA_SEARCH = "EXA_SEARCH",
  SERPER_SEARCH = "SERPER_SEARCH",
  NEWSAPI = "NEWSAPI",
  GDELT = "GDELT",
  HACKERNEWS = "HACKERNEWS",
  ARXIV = "ARXIV",
  ABOUTAMAZON = "ABOUTAMAZON",
  FIRECRAWL = "FIRECRAWL",
  RSSHUB = "RSSHUB",
  TWITTER = "TWITTER",
}